// packages/react/src/components/PageNavigation/utils/type.ts
export type PageNavigationStructureType = {
    displayedRange: number[];
    nextPageNumber: number;
    previousPageNumber: number;
    currentPageNumber: number;
    firstPageArrowEnabled: boolean;
    lastPageArrowEnabled: boolean;
    nextPageArrowEnabled: boolean;
    previousPageArrowEnabled: boolean;
};